Convert all SIRI tests to module tests
